+++ integration-ts/examples/flow/startWith-moveUntil-dropTo.ts
@@ -1,20 +1,15 @@
 import {Streams, F, C,Option, N, SingleParser} from '@robusta/trash'
 import {assertFalse, assertTrue} from '../../assert';
 
-function day() {
-    return C.stringIn(['MONDAY', 'TUESDAY', 'WEDNESDAY', 'THURSDAY', 'FRIDAY', 'SATURDAY', 'SUNDAY']);
-}
-
-function a(){
-    return C.char('a');
-}
-
-const string = 'Xabx';
+const string = 'The quick brown fox jumps over the lazy dog';
 
 
 
 function combinator() {
-    return F.not(day()).then(F.not(a())).then(F.not(day()));
+    return F.startWith('hello: ')
+        .then(F.moveUntil('brown'))
+        .then(C.string('brown'))
+        .then(F.dropTo(F.eos()));
 }
 
 let stream = Streams.ofString(string);
@@ -22,4 +17,5 @@ let parsing = combinator().parse(stream);
 
 assertFalse(parsing.isAccepted());
 assertTrue(parsing.offset === 2);
+assertTrue(parsing.value.join('') === 'hello: The quick brown');